backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.32
alembic>=1.13.2
python-multipart>=0.0.9
pydantic>=2.8.2
pydantic-settings>=2.3.4
argon2-cffi>=23.1.0
tzdata>=2025.2
openpyxl>=3.1.5
httpx>=0.28.0
